Feat #43 : Signup logic 변경
- 회원가입 후, 메인 페이지에서 모달 형식으로 트랙 수정할 예정
