+++ src/Utils/Stores/config.ts
@@ -1,0 +1,7 @@
+import theme from '@/Styles/theme';
+
+export const toastConfig = [
+  { type: 'success', icon: 'mdi:success-circle', color: `${theme.color.good}` },
+  { type: 'warn', icon: 'material-symbols:warning-rounded', color: `${theme.color.ajouGold}` },
+  { type: 'error', icon: 'ic:baseline-error', color: `${theme.color.bad}` },
+];
